+++ collector-amazon-ivs/src/androidTest/java/com/bitmovin/analytics/amazon/ivs/IvsPlayerConstants.kt
@@ -1,7 +1,0 @@
-package com.bitmovin.analytics.amazon.ivs
-
-import com.bitmovin.analytics.test.utils.PlayerInfo
-
-object IvsPlayerConstants {
-    val playerInfo = PlayerInfo("amazonivs", "amazonivs-1.38.0", "Android:AmazonIVS")
-}