+++ collector-amazon-ivs/src/main/java/com/bitmovin/analytics/amazon/ivs/Utils.kt
@@ -1,9 +1,0 @@
-package com.bitmovin.analytics.amazon.ivs
-
-internal class Utils {
-    companion object {
-        fun isPlaybackLive(duration: Long): Boolean {
-            return duration == -1L
-        }
-    }
-}