+++ collector-exoplayer/src/androidTest/java/com/bitmovin/analytics/exoplayer/ExoplayerConstants.kt
@@ -1,7 +1,0 @@
-package com.bitmovin.analytics.exoplayer
-
-import com.bitmovin.analytics.test.utils.PlayerInfo
-
-object ExoplayerConstants {
-    val playerInfo = PlayerInfo("exoplayer", "exoplayer-2.19.1", "Android:Exoplayer")
-}