+++ publish.sh
@@ -78,9 +78,7 @@ echo ""
 echo "Artifacts to publish:"
 echo "  - com.bitmovin.analytics:collector:$VERSION (:collector project)"
 echo "  - com.bitmovin.analytics:collector-bitmovin-player:$VERSION (:collector-bitmovin-player project)"
-echo "  - com.bitmovin.analytics:collector-exoplayer:$VERSION (:collector-exoplayer project)"
 echo "  - com.bitmovin.analytics:collector-media3-exoplayer:$VERSION (:collector-media3-exoplayer project)"
-echo "  - com.bitmovin.analytics:collector-amazon-ivs:$VERSION (:collector-amazon-ivs project)"
 echo "  - com.bitmovin.analytics:collector-theoplayer:$VERSION (:collector-theoplayer project)"
 echo "\nAre all tokens, artifacts and versions correct ?"
 read -p "(Press enter to continue)"
@@ -116,22 +114,6 @@ echo "\n:collector-bitmovin-player project build and publishing..."
 ./gradlew -DdevelopLocal=false -Dversion="$VERSION" :collector-bitmovin-player:artifactoryPublish || exit
 ./gradlew -DdevelopLocal=false -Dversion="$VERSION" :collector-bitmovin-player:publishToMavenLocal || exit
 echo "\n:collector-bitmovin-player project built and published!"
-
-echo "\n:collector-exoplayer project build and publishing..."
-./gradlew -DdevelopLocal=false -Dversion="$VERSION" :collector-exoplayer:clean || exit
-./gradlew -DdevelopLocal=false -Dversion="$VERSION" :collector-exoplayer:build || exit
-./gradlew -DdevelopLocal=false -Dversion="$VERSION" :collector-exoplayer:assembleRelease || exit
-./gradlew -DdevelopLocal=false -Dversion="$VERSION" :collector-exoplayer:artifactoryPublish || exit
-./gradlew -DdevelopLocal=false -Dversion="$VERSION" :collector-exoplayer:publishToMavenLocal || exit
-echo "\n:collector-exoplayer project built and published!"
-
-echo "\n:collector-amazon-ivs project build and publishing..."
-./gradlew -DdevelopLocal=false -Dversion="$VERSION" :collector-amazon-ivs:clean || exit
-./gradlew -DdevelopLocal=false -Dversion="$VERSION" :collector-amazon-ivs:build || exit
-./gradlew -DdevelopLocal=false -Dversion="$VERSION" :collector-amazon-ivs:assembleRelease || exit
-./gradlew -DdevelopLocal=false -Dversion="$VERSION" :collector-amazon-ivs:artifactoryPublish || exit
-./gradlew -DdevelopLocal=false -Dversion="$VERSION" :collector-amazon-ivs:publishToMavenLocal || exit
-echo "\n:collector-amazon-ivs project built and published!"
 
 echo "\n:collector-media3-exoplayer project build and publishing..."
 ./gradlew -DdevelopLocal=false -Dversion="$VERSION" :collector-media3-exoplayer:clean || exit
@@ -185,8 +167,6 @@ echo "Copying artifacts from libs-release-local to public-releases in jfrog ..."
 
 curl -H "Content-Type: application/json" -X POST -u ${artifactoryUser}:${artifactoryPassword} "https://bitmovin.jfrog.io/bitmovin/api/copy/libs-release-local/com/bitmovin/analytics/collector/${VERSION}?to=/public-releases/com/bitmovin/analytics/collector/${VERSION}"
 curl -H "Content-Type: application/json" -X POST -u ${artifactoryUser}:${artifactoryPassword} "https://bitmovin.jfrog.io/bitmovin/api/copy/libs-release-local/com/bitmovin/analytics/collector-bitmovin-player/${VERSION}?to=/public-releases/com/bitmovin/analytics/collector-bitmovin-player/${VERSION}"
-curl -H "Content-Type: application/json" -X POST -u ${artifactoryUser}:${artifactoryPassword} "https://bitmovin.jfrog.io/bitmovin/api/copy/libs-release-local/com/bitmovin/analytics/collector-exoplayer/${VERSION}?to=/public-releases/com/bitmovin/analytics/collector-exoplayer/${VERSION}"
-curl -H "Content-Type: application/json" -X POST -u ${artifactoryUser}:${artifactoryPassword} "https://bitmovin.jfrog.io/bitmovin/api/copy/libs-release-local/com/bitmovin/analytics/collector-amazon-ivs/${VERSION}?to=/public-releases/com/bitmovin/analytics/collector-amazon-ivs/${VERSION}"
 curl -H "Content-Type: application/json" -X POST -u ${artifactoryUser}:${artifactoryPassword} "https://bitmovin.jfrog.io/bitmovin/api/copy/libs-release-local/com/bitmovin/analytics/collector-media3-exoplayer/${VERSION}?to=/public-releases/com/bitmovin/analytics/collector-media3-exoplayer/${VERSION}"
 curl -H "Content-Type: application/json" -X POST -u ${artifactoryUser}:${artifactoryPassword} "https://bitmovin.jfrog.io/bitmovin/api/copy/libs-release-local/com/bitmovin/analytics/collector-theoplayer/${VERSION}?to=/public-releases/com/bitmovin/analytics/collector-theoplayer/${VERSION}"
 
@@ -195,8 +175,6 @@ echo "\nCopied artifacts to public jfrog repo."
 
 echo "\nNotifying bitmovin api about new release..."
 notifyApi "android-bitmovin" $VERSION "collector-bitmovin-player"
-notifyApi "android-exo" $VERSION "collector-exoplayer"
-notifyApi "android-amazon-ivs" $VERSION "collector-amazon-ivs"
 notifyApi "android-media3-exo" $VERSION "collector-media3-exoplayer"
 
 ## TODO: this needs to be added once we have a stable version